+++ php/tests/Application/Command/LocalizeTypeTest/LocalizeType.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace A
+{
+
+}
+
+namespace B
+{
+
+use DateTime;
+use C\D;
+use C\D\E;
+use F\G\H;
+use F\G;
+
+}